backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
pymupdf>=1.24.0
selenium>=4.17.0
opencc-python-reimplemented>=0.1.7
python-multipart>=0.0.7
aiofiles>=23.2.1
websockets>=12.0
httpx>=0.27.0
